+++ K4AUnityBTPlugin/K4AUnityBTPluginTest/K4AUnityBTPluginTest.cpp
@@ -25,7 +25,7 @@ namespace K4AUnityBTPluginTest
 			KinectBodyTracker::Body bodies[K4ABT_MAX_BODY];
 			K4ABT_GetBody(bodies, numBodies);
 
- 			for (auto i = 0; i < numBodies; i++)
+ 			for (auto i = 0; i < (int)numBodies; i++)
 			{
 				for (auto j = 0; j < K4ABT_JOINT_COUNT; j++)
 				{
